Initial checking
Implementation of TCPH Q1 without Group by or date filter, these
features will be available later on.
Goal is to measure performance gains from writing hand optimized code a
way of simulating code generations (best case scenario) and evaluate
SIMD and try to incorporate it as much as possible in solving the query.